+++ test/__mocks__/mongodb.js
@@ -1,0 +1,7 @@
+export const MongoClient = {
+  connect: vi.fn(),
+};
+
+export default {
+  MongoClient,
+};